Move Some Operations From OrderedSet To SetOpsForOrderedSets
This will allow sharing them with NonEmptySet In the future.
